Switch to IGitHubWrapper
Проект GitHub было решено удалить и переместить код из него в Model. Так
я подготавливаю кодовую базу к добовлению IGitHubFacade. Он будет
переводить понятия Гитхаба в понятия Гиры.
