+++ Source/Gira.sln
@@ -9,8 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ModelTests", "ModelTests\Mo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebAPI", "WebAPI\WebAPI.csproj", "{303DD3FB-5F12-436C-825F-232D26A5D9BF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GitHub", "GitHub\GitHub.csproj", "{48F4DBB5-ED76-411D-B0F1-E74790B99847}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -57,18 +55,6 @@ Global
 		{303DD3FB-5F12-436C-825F-232D26A5D9BF}.Release|x64.Build.0 = Release|Any CPU
 		{303DD3FB-5F12-436C-825F-232D26A5D9BF}.Release|x86.ActiveCfg = Release|Any CPU
 		{303DD3FB-5F12-436C-825F-232D26A5D9BF}.Release|x86.Build.0 = Release|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Debug|x64.Build.0 = Debug|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Debug|x86.Build.0 = Debug|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Release|Any CPU.Build.0 = Release|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Release|x64.ActiveCfg = Release|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Release|x64.Build.0 = Release|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Release|x86.ActiveCfg = Release|Any CPU
-		{48F4DBB5-ED76-411D-B0F1-E74790B99847}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
