--- Source/GitHub/GitHub.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="GitHubJwt" Version="0.0.4" />
-    <PackageReference Include="Octokit" Version="0.50.0" />
-    <PackageReference Include="Octokit.GraphQL" Version="0.1.7-beta" />
-  </ItemGroup>
-
-</Project>
