--- Source/GitHub/Installations/Installation.cs
@@ -1,20 +1,0 @@
-﻿namespace GitHub
-{
-	public class Installation
-	{
-		public long Id { get; private set; }
-
-		public Owner Owner { get; private set; }
-
-		public Installation(long id, Owner owner)
-		{
-			Id = id;
-			Owner = owner;
-		}
-
-		internal static Installation BuildFrom(Octokit.Installation installation, Owner owner)
-		{
-			return new Installation(installation.Id, owner);
-		}
-	}
-}
