--- Source/GitHub/Owners/Owner.cs
@@ -1,17 +1,0 @@
-﻿namespace GitHub
-{
-	public class Owner
-	{
-		public string Login { get; private set; }
-
-		public Owner(string login)
-		{
-			Login = login;
-		}
-
-		internal static Owner BuildFrom(Octokit.Account account)
-		{
-			return new Owner(account.Login);
-		}
-	}
-}
