+++ Source/Model/GitHubWrapper/Project.cs
@@ -1,0 +1,15 @@
+﻿namespace Model
+{
+	public class Project
+	{
+		public int Id { get; private set; }
+
+		public string Name { get; private set; }
+
+		public Project(int id, string name)
+		{
+			Id = id;
+			Name = name;
+		}
+	}
+}
