+++ Source/Model/Repos/RepoSettings.cs
@@ -1,5 +1,4 @@
-﻿using GitHub;
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace Model
