+++ Source/Model/Sprints/SprintsDataSource.cs
@@ -1,5 +1,4 @@
-﻿using GitHub;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 
 namespace Model
